Create lc-overview-decommission-notes
Per 2023-09-21 LC-CAC decision, I am working to decommission the LC-Overview lesson. 

I need a community-sponsored place to keep track of notes and actions so that I can share notes with people and leave notes that I can refer back to.

This repo and this folder seem like a good location for that.